+++ deps.ts
@@ -1,20 +1,20 @@
 import Denomander from "https://deno.land/x/denomander@0.5.1/mod.ts";
 import stdConfig from "https://deno.land/x/nessie/nessie.config.ts";
 
-export { Denomander, stdConfig };
+export {
+  assert,
+  assertArrayContains,
+  assertEquals,
+} from "https://deno.land/std@v0.51.0/testing/asserts.ts";
 export {
   Client as MySQLClient,
   ClientConfig,
 } from "https://deno.land/x/mysql@2.0.0/mod.ts";
-export { Client as PGClient } from "https://deno.land/x/postgres@v0.3.11/mod.ts";
 export { IConnectionParams } from "https://deno.land/x/postgres@v0.3.11/connection_params.ts";
+export { Client as PGClient } from "https://deno.land/x/postgres@v0.3.11/mod.ts";
 export {
   DB as SQLiteClient,
-  save,
   open,
+  save,
 } from "https://deno.land/x/sqlite@v1.0.0/mod.ts";
-export {
-  assertEquals,
-  assert,
-  assertArrayContains,
-} from "https://deno.land/std@v0.51.0/testing/asserts.ts";
+export { Denomander, stdConfig };
